set ts paths and move code to src
