+++ packages/covector/tsconfig.json
@@ -1,8 +1,16 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
+    "baseUrl": ".",
     "outDir": "dist",
-    "rootDir": "src"
+    "rootDir": "./src"
   },
-  "include": ["./src/**/*.ts"]
+  "references": [
+    { "path": "../files" },
+    { "path": "../assemble" },
+    { "path": "../apply" },
+    { "path": "../changelog" },
+    { "path": "../command" }
+  ],
+  "include": ["src/**/*.ts"]
 }
