+++ tsconfig.packages.json
@@ -2,10 +2,10 @@
   "files": [],
   "references": [
     { "path": "./packages/files" },
+    { "path": "./packages/command" },
+    { "path": "./packages/changelog" },
     { "path": "./packages/assemble" },
     { "path": "./packages/apply" },
-    { "path": "./packages/changelog" },
-    { "path": "./packages/command" },
     { "path": "./packages/covector" }
   ]
 }
